feat(core): extract SignerMultisigCkbBase for Omnilock multisig
Extract shared multisig logic into abstract SignerMultisigCkbBase with
encodeWitnessLock/decodeWitnessLock template methods. CKB multisig and
Omnilock multisig are thin subclasses differing only in witness encoding.

MultisigCkbWitness entity moved to its own module. OmniLockWitnessLock
molecule encoding added as standalone utility.
