+++ packages/core/src/signer/ckb/index.ts
@@ -1,7 +1,12 @@
+export * from "./multisigCkbWitness.js";
+export * from "./omniLockWitnessLock.js";
 export * from "./secp256k1Signing.js";
 export * from "./signerCkbPrivateKey.js";
 export * from "./signerCkbPublicKey.js";
 export * from "./signerCkbScriptReadonly.js";
+export * from "./signerMultisigCkbBase.js";
 export * from "./signerMultisigCkbPrivateKey.js";
 export * from "./signerMultisigCkbReadonly.js";
+export * from "./signerMultisigOmniLockPrivateKey.js";
+export * from "./signerMultisigOmniLockReadonly.js";
 export * from "./verifyJoyId.js";
